refactor (Model과 View, Controller로 역할 구분 및 패키지 구분)
Model과 View, Controller로 역할 구분 및 패키지 구분
- Model과 View 패키지 생성 및 각 클래스 파일 이동
- RacingGame의 역할을 Controller로 변경하고 Validation 로직을 따로 클래스로 분리
- Validation 로직에 대한 테스트코드 생성
