+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,9 +1,12 @@
 package racingcar;
 
+import racingcar.domain.RandomNumberGenerator;
+import racingcar.view.RacingGameUI;
+
 public class RacingGameApplication {
 
 	public static void main(String[] args) {
-		RacingGame game = new RacingGame(
+		RacingGameController game = new RacingGameController(
 			new RacingGameUI(),
 			new RandomNumberGenerator()
 		);
