+++ src/main/java/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Objects;
 import racingcar.dto.CarState;
