+++ src/main/java/racingcar/domain/ProceedLogic.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class ProceedLogic {
 	private final NumberGenerator numberGenerator;
