+++ src/test/java/racingcar/CarTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import racingcar.domain.Car;
+
 public class CarTest {
 
 	@Test
