+++ src/test/java/racingcar/ProceedLogicTest.java
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import racingcar.domain.ProceedLogic;
+
 @DisplayName("전진 여부를 판단하는 클래스 테스트")
 public class ProceedLogicTest {
